feat: export static site
- Add npm run scripts to `export` and `serve` the static site as per https://github.com/glifio/wallet/issues/513
- Use `serve` as it's the hip vercel flavour static server.
- It blends! previously it did not! I did nothing but try again!

I was able to use the local static site to connect to a Ledger and view some accounts.

So all aboard the content-addressing train 🚂🔗🔐✨ ...any reason not to host glif on IPFS? I can PR a CI job to build and pin and dnslink the static site if it is acceptable.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>