fix issue where this was opening in a new window which was breaking the functionality
It’s weird that this just started breaking because looking at all the
relevant files, none of them have been changed recently. So I'm not
sure when or how this breaking change happened. This probably isn't the
correct fix. So we probably want to figure out how to not have
target=‘_blank’ as a default.